combined login.tsx with page.tsx in /src folder
