+++ src/app/LoginPage.module.css
@@ -1,0 +1,90 @@
+@import-normalize;
+
+.container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    min-height: 100vh;
+    margin: 0;
+    padding: 0;
+}
+
+.background {
+    position: relative;
+    width: 100%;
+    height: 100vh;
+    overflow: hidden;
+}
+
+.stage {
+    position: absolute;
+    top: 0;
+    left: 0;
+
+    display: block;
+    z-index: 1;
+}
+
+.login {
+    /* Set positioning of login box */
+    position: absolute;
+    top: 50%;
+    left: 12.5%;
+    /* transform: translateX(50%); */
+    transform: translateY(-50%);
+    z-index: 2;
+
+    /* Sizing options */
+    width: 37.5%;
+    height: 70%;
+    /* max-width: 400px; */
+
+    /* Formatting options */
+    background-color: rgba(255, 255, 255, 0.95);
+    padding: 20px;
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+
+    /* Flexbox/display options */
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    box-sizing: border-box;
+}
+
+.login>* {
+    font: Poppins;
+}
+
+.login_title {
+    margin-bottom: 50px;
+    font-size: 2.5em;
+}
+
+.login_input {
+    width: 80%;
+    margin-bottom: 20px;
+    padding: 10px;
+    border: 1px solid #ccc;
+    border-radius: 5px;
+}
+
+.login_button {
+    width: 30%;
+    padding: 10px;
+    margin-top: 20px;
+    background-color: rgb(0, 0, 0);
+    color: rgb(255, 255, 255);
+    border: none;
+    border-radius: 10px;
+    cursor: pointer;
+}
+
+.forgot_password {
+    text-align: right;
+}
+
+
+.logo {
+    width: 50%;
+    height: auto;
+}
