--- src/app/login.module.css
@@ -1,161 +1,0 @@
-@import-normalize;
-
-.background {
-    position: relative;
-    width: 100%;
-    height: 100vh;
-    overflow: hidden;
-}
-
-.stage {
-    position: absolute;
-    top: 0;
-    left: 0;
-
-    display: block;
-    z-index: 1;
-}
-
-.login {
-    /* Set positioning of login box */
-    position: absolute;
-    top: 50%;
-    left: 12.5%;
-    /* transform: translateX(50%); */
-    transform: translateY(-50%);
-    z-index: 2;
-
-    /* Sizing options */
-    width: 37.5%;
-    height: 70%;
-    /* max-width: 400px; */
-
-    /* Formatting options */
-    background-color: rgba(255, 255, 255, 0.95);
-    padding: 20px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-
-    /* Flexbox/display options */
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    box-sizing: border-box;
-}
-
-.login > * {
-    font: Poppins;
-}
-
-.login_title {
-    margin-bottom: 50px;
-    font-size: 2.5em;
-}
-
-.login_input {
-    width: 80%;
-    margin-bottom: 20px;
-    padding: 10px;
-    border: 1px solid #ccc;
-    border-radius: 5px;
-}
-
-.login_button {
-    width: 30%;
-    padding: 10px;
-    margin-top: 20px;
-    background-color: rgb(0,0,0);
-    color: rgb(255,255,255);
-    border: none;
-    border-radius: 10px;
-    cursor: pointer;
-}
-
-.forgot_password {
-    text-align: right;
-}
-
-
-.logo {
-    width: 50%;
-    height: auto;
-}
-
-
-
-/* .background {
-    margin: 0px;
-    padding: 0px;
-
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    display: flex;
-    justify-content: center;
-    height: 100%;
-    position: absolute;
-    left: 0;
-    width: 100%;
-    overflow: hidden;
-}
-
-.login_container {
-    display: flex;
-
-    align-items: center;
-    justify-content: space-between;
-    height: 70%;
-    width:  70%;
-}
-
-.login_subcontainer {
-    flex-grow: 1;
-    background-color: aqua;
-
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-    width: 50%;
-    height: 100%;
-}
-
-.login_filler {
-    flex-grow: 1;
-
-    width: 50%;
-    height: 100%;
-}
-
-.title_container {
-    display: flex;
-    flex-direction: column;
-    font-size: 64px;
-    font-weight: bolder;
-    align-items: center;
-    justify-content: center;
-}
-
-.input_container {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    justify-content: center; 
-
-    padding: 8px;
-}
-
-.input_container > .errorLabel {
-  color: red;
-  font-size: 12px;
-}
-
-
-.input_box {
-    height: 40px;
-    width: 400px;
-    font-size: large;
-    border-radius: 8px;
-    border: 1px solid grey;
-    padding-left: 8px;
-} */
